backend/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
WeasyPrint==52.5
drf-base64==2.0
psycopg-binary==2.9.6
python-dotenv==1.0.0
